Renaming mostly works
A second container (such as the basic inventory) needs to be opened before opening of the sack container as the sack name will otherwise not change (though the text field does use the new name ugh). Likely will be solved via IContainerListener. Item is properly renamed and name shows up in nameField otherwise